// webclient/src/gen/rendezvous.ts
export enum ConnType {
  DEFAULT_CONN = 0,
  FILE_TRANSFER = 1,
  PORT_FORWARD = 2,
  RDP = 3,
  VIEW_CAMERA = 4,
  TERMINAL = 5,
}

export enum NatType {
  UNKNOWN_NAT = 0,
  ASYMMETRIC = 1,
  SYMMETRIC = 2,
}

export interface RegisterPeer {
  id: string;
  serial: number;
}

export interface RegisterPeerResponse {
  request_pk: boolean;
}

export interface PunchHoleRequest {
  id: string;
  nat_type: NatType;
  licence_key: string;
  conn_type: ConnType;
  token: string;
  version: string;
  udp_port: number;
  force_relay: boolean;
  upnp_port: number;
  socket_addr_v6: Uint8Array;
  switch_code: string;
}
// ...
export enum PunchHoleResponse_Failure {
  ID_NOT_EXIST = 0,
  OFFLINE = 2,
  LICENSE_MISMATCH = 3,
  LICENSE_OVERUSE = 4,
}
// ...
export interface OnlineRequest {
  id: string;
  peers: string[];
}
// ...
export interface RendezvousMessage {
  union?:
    | { $case: "register_peer"; register_peer: RegisterPeer }
    | { $case: "register_peer_response"; register_peer_response: RegisterPeerResponse }
    | { $case: "punch_hole_request"; punch_hole_request: PunchHoleRequest }
    | { $case: "punch_hole"; punch_hole: PunchHole }
    | { $case: "punch_hole_sent"; punch_hole_sent: PunchHoleSent }
    | { $case: "punch_hole_response"; punch_hole_response: PunchHoleResponse }
    | { $case: "fetch_local_addr"; fetch_local_addr: FetchLocalAddr }
    | { $case: "local_addr"; local_addr: LocalAddr }
    | { $case: "configure_update"; configure_update: ConfigUpdate }
    | { $case: "register_pk"; register_pk: RegisterPk }
    | { $case: "register_pk_response"; register_pk_response: RegisterPkResponse }
    | { $case: "software_update"; software_update: SoftwareUpdate }
    | { $case: "request_relay"; request_relay: RequestRelay }
    | { $case: "relay_response"; relay_response: RelayResponse }
    | { $case: "test_nat_request"; test_nat_request: TestNatRequest }
    | { $case: "test_nat_response"; test_nat_response: TestNatResponse }
    | { $case: "peer_discovery"; peer_discovery: PeerDiscovery }
    | { $case: "online_request"; online_request: OnlineRequest }
    | { $case: "online_response"; online_response: OnlineResponse }
    | { $case: "key_exchange"; key_exchange: KeyExchange }
    | { $case: "hc"; hc: HealthCheck }
    | { $case: "http_proxy_request"; http_proxy_request: HttpProxyRequest }
    | { $case: "http_proxy_response"; http_proxy_response: HttpProxyResponse }
    | undefined;
}

function createBaseRegisterPeer(): RegisterPeer {
  return { id: "", serial: 0 };
}

export const RegisterPeer: MessageFns<RegisterPeer> = {
  encode(message: RegisterPeer, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.serial !== 0) {
      writer.uint32(16).int32(message.serial);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RegisterPeer {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRegisterPeer();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.serial = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<RegisterPeer>, I>>(base?: I): RegisterPeer {
    return RegisterPeer.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RegisterPeer>, I>>(object: I): RegisterPeer {
    const message = createBaseRegisterPeer();
    message.id = object.id ?? "";
    message.serial = object.serial ?? 0;
    return message;
  },
};

function createBaseRegisterPeerResponse(): RegisterPeerResponse {
  return { request_pk: false };
}

export const RegisterPeerResponse: MessageFns<RegisterPeerResponse> = {
  encode(message: RegisterPeerResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.request_pk !== false) {
      writer.uint32(16).bool(message.request_pk);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): RegisterPeerResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRegisterPeerResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.request_pk = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<RegisterPeerResponse>, I>>(base?: I): RegisterPeerResponse {
    return RegisterPeerResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<RegisterPeerResponse>, I>>(object: I): RegisterPeerResponse {
    const message = createBaseRegisterPeerResponse();
    message.request_pk = object.request_pk ?? false;
    return message;
  },
};

function createBasePunchHoleRequest(): PunchHoleRequest {
  return {
    id: "",
    nat_type: 0,
    licence_key: "",
    conn_type: 0,
    token: "",
    version: "",
    udp_port: 0,
    force_relay: false,
    upnp_port: 0,
    socket_addr_v6: new Uint8Array(0),
    switch_code: "",
  };
}

export const PunchHoleRequest: MessageFns<PunchHoleRequest> = {
  encode(message: PunchHoleRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    if (message.nat_type !== 0) {
      writer.uint32(16).int32(message.nat_type);
    }
    if (message.licence_key !== "") {
      writer.uint32(26).string(message.licence_key);
    }
    if (message.conn_type !== 0) {
      writer.uint32(32).int32(message.conn_type);
    }
    if (message.token !== "") {
      writer.uint32(42).string(message.token);
    }
    if (message.version !== "") {
      writer.uint32(50).string(message.version);
    }
    if (message.udp_port !== 0) {
      writer.uint32(56).int32(message.udp_port);
    }
    if (message.force_relay !== false) {
      writer.uint32(64).bool(message.force_relay);
    }
    if (message.upnp_port !== 0) {
      writer.uint32(72).int32(message.upnp_port);
    }
    if (message.socket_addr_v6.length !== 0) {
      writer.uint32(82).bytes(message.socket_addr_v6);
    }
    if (message.switch_code !== "") {
      writer.uint32(90).string(message.switch_code);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PunchHoleRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePunchHoleRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.nat_type = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.licence_key = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.conn_type = reader.int32() as any;
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.token = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.version = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 56) {
            break;
          }

          message.udp_port = reader.int32();
          continue;
        }
        case 8: {
          if (tag !== 64) {
            break;
          }

          message.force_relay = reader.bool();
          continue;
        }
        case 9: {
          if (tag !== 72) {
            break;
          }

          message.upnp_port = reader.int32();
          continue;
        }
        case 10: {
          if (tag !== 82) {
            break;
          }

          message.socket_addr_v6 = reader.bytes();
          continue;
        }
        case 11: {
          if (tag !== 90) {
            break;
          }

          message.switch_code = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<PunchHoleRequest>, I>>(base?: I): PunchHoleRequest {
    return PunchHoleRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PunchHoleRequest>, I>>(object: I): PunchHoleRequest {
    const message = createBasePunchHoleRequest();
    message.id = object.id ?? "";
    message.nat_type = object.nat_type ?? 0;
    message.licence_key = object.licence_key ?? "";
    message.conn_type = object.conn_type ?? 0;
    message.token = object.token ?? "";
    message.version = object.version ?? "";
    message.udp_port = object.udp_port ?? 0;
    message.force_relay = object.force_relay ?? false;
    message.upnp_port = object.upnp_port ?? 0;
    message.socket_addr_v6 = object.socket_addr_v6 ?? new Uint8Array(0);
    message.switch_code = object.switch_code ?? "";
    return message;
  },
};
// ...
export const OnlineRequest: MessageFns<OnlineRequest> = {
  encode(message: OnlineRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== "") {
      writer.uint32(10).string(message.id);
    }
    for (const v of message.peers) {
      writer.uint32(18).string(v!);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): OnlineRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseOnlineRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.peers.push(reader.string());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<OnlineRequest>, I>>(base?: I): OnlineRequest {
    return OnlineRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<OnlineRequest>, I>>(object: I): OnlineRequest {
    const message = createBaseOnlineRequest();
    message.id = object.id ?? "";
    message.peers = object.peers?.map((e) => e) || [];
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends bigint ? string | number | bigint
  : T extends Builtin ? T : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}